Move variable display into separate file.
